refactor: extract sanitize tests into separate file
